server/requirements.txt
SQLAlchemy~=2.0.38
bcrypt~=4.2.1
PyJWT~=2.10.1
fastapi~=0.115.8
cloudinary~=1.42.2
python-dotenv~=1.0.1
pydantic~=2.10.6